pyproject.toml
[project]
name = "zalo-decrypt-backup"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "pycryptodome>=3.21.0",
    "pywin32>=308",
    "httpx>=0.27.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []
